REFACTOR: click outside functionality into hook
