+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+import useClickOutside from './useClickOutside.hook';
+
+export { useClickOutside };
